Get latest box version via HTTP GET instead of "vagrant box outdated"  * Moved constants to constants package  * Moved file download from mebroutines/install to network package  * Moved vagrant box version dection stuff from mebroutines    to boxversion package
