+++ src/naksu/mebroutines/install/install.go
@@ -1,20 +1,17 @@
 package install
 
 import (
-	"errors"
 	"fmt"
-	"io"
+	"naksu/constants"
 	"naksu/mebroutines"
+	"naksu/network"
 	"naksu/progress"
 	"naksu/xlate"
-	"net/http"
 	"os"
 )
 
 // GetServer downloads vagrantfile and starts server
 func GetServer(newVagrantfilePath string) {
-	const VagrantURL = "http://static.abitti.fi/usbimg/qa/vagrant/Vagrantfile"
-
 	// Create ~/ktp if missing
 	progress.TranslateAndSetMessage("Creating ~/ktp")
 	var ktpPath = createKtpDir()
@@ -33,7 +30,7 @@ func GetServer(newVagrantfilePath string) {
 		abittiVagrantfilePath := vagrantfilePath + ".abitti"
 
 		progress.TranslateAndSetMessage("Downloading Abitti Vagrantfile")
-		errDownload := downloadFile(VagrantURL, abittiVagrantfilePath)
+		errDownload := network.DownloadFile(constants.AbittiVagrantURL, abittiVagrantfilePath)
 		if errDownload != nil {
 			mebroutines.LogDebug(fmt.Sprintf("Download failed: %v", errDownload))
 			mebroutines.ShowWarningMessage(xlate.Get("Could not update Abitti stickless server. Please check your network connection."))
@@ -117,29 +114,4 @@ func removeVagrantfile(vagrantfilePath string) {
 	if err != nil {
 		mebroutines.ShowWarningMessage(fmt.Sprintf(xlate.Get("Failed to rename %s to %s"), vagrantfilePath, vagrantfilePath+".bak"))
 	}
-}
-
-func downloadFile(url string, filepath string) error {
-	mebroutines.LogDebug(fmt.Sprintf("Starting download from URL %s to file %s", url, filepath))
-
-	out, err1 := os.Create(filepath)
-	if err1 != nil {
-		return errors.New("failed to create file")
-	}
-	defer mebroutines.Close(out)
-
-	/* #nosec */
-	resp, err2 := http.Get(url)
-	if err2 != nil {
-		return errors.New("failed to retrieve file")
-	}
-	defer mebroutines.Close(resp.Body)
-
-	_, err3 := io.Copy(out, resp.Body)
-	if err3 != nil {
-		return errors.New("failed to copy body")
-	}
-
-	mebroutines.LogDebug(fmt.Sprintf("Finished download from URL %s to file %s", url, filepath))
-	return nil
 }
